restructure app to make it extendable
