+++ backend/main.py
@@ -1,13 +1,8 @@
 import uvicorn
-from fastapi import Depends, FastAPI
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from sqlalchemy.orm import Session
 
-from backend import service, models
-from backend.database import dbschema
-from backend.database.config import SessionLocal, engine
-
-dbschema.Base.metadata.create_all(bind=engine)
+from backend.endpoints.api import api_router
 
 app = FastAPI()
 
@@ -24,25 +19,6 @@ app.add_middleware(
 )
 
 
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
-@app.post("/articles/create/")
-def create_article(article: models.ArticleCreate, db: Session = Depends(get_db)):
-    return service.create_article(db=db, article=article)
-
-
-@app.get("/articles/", response_model=list[models.Article])
-def get_articles(db: Session = Depends(get_db)):
-    articles = service.get_articles(db)
-    return articles
-
-
 # Below method is being used as a sanity check when debugging to make sure the debugger is actually working
 @app.get("/")
 def root():
@@ -50,6 +26,8 @@ def root():
     b = "b" + a
     return {"hello world": b}
 
+
+app.include_router(api_router)
 
 if __name__ == "__main__":
     uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True)
